+++ framework-core/framework-net/src/main/java/org/smart/framework/net/router/annotation/Module.java
@@ -12,5 +12,5 @@ public @interface Module {
 	 * 模块映射
 	 * @return
 	 */
-	int id();
+	byte id();
 }
